backend/requirements.txt
fastapi
uvicorn[standard]
pymupdf4llm
chromadb
langchain
langchain-openai
langchain-community
openai
python-dotenv
pydantic
